requirements.txt
streamlit
yt-dlp
moviepy
torch
speechbrain
ffmpeg-python
